refactor: 누락된 코드 컨벤션 수정
- 파라미터 final 추가
- NotBlank와 함께 사용되던 NotNull 삭제
- dto 패키지 추가
- organization id 순서 변경
