--- server/src/main/java/com/ahmadda/application/OrganizationCreateRequest.java
@@ -1,10 +1,0 @@
-package com.ahmadda.application;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-
-public record OrganizationCreateRequest(@NotBlank @NotNull String name,
-                                        @NotBlank @NotNull String description,
-                                        @NotNull String imageUrl) {
-
-}
