+++ server/src/main/java/com/ahmadda/application/OrganizationService.java
@@ -1,5 +1,6 @@
 package com.ahmadda.application;
 
+import com.ahmadda.application.dto.OrganizationCreateRequest;
 import com.ahmadda.application.exception.NotFoundException;
 import com.ahmadda.domain.Organization;
 import com.ahmadda.domain.OrganizationRepository;
