+++ server/src/main/java/com/ahmadda/application/dto/OrganizationCreateRequest.java
@@ -1,0 +1,10 @@
+package com.ahmadda.application.dto;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+
+public record OrganizationCreateRequest(@NotBlank String name,
+                                        @NotBlank String description,
+                                        @NotNull String imageUrl) {
+
+}
