+++ server/src/main/java/com/ahmadda/presentation/OrganizationController.java
@@ -1,8 +1,8 @@
 package com.ahmadda.presentation;
 
-import com.ahmadda.application.OrganizationCreateRequest;
-import com.ahmadda.application.OrganizationReadResponse;
 import com.ahmadda.application.OrganizationService;
+import com.ahmadda.application.dto.OrganizationCreateRequest;
+import com.ahmadda.application.dto.OrganizationReadResponse;
 import com.ahmadda.domain.Organization;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
